Created boilerplate for package.json file with next scripts. Renamed peter.ts to constructors.ts to reflect its content.
