+++ CHANGELOG.md
@@ -2,8 +2,16 @@
 
 All notable changes to the "upnext" extension will be documented in this file.
 
-Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how to structure this file.
+## [0.0.2] - 2021-04-21
 
-## [Unreleased]
+### Add
 
-- Initial release
+- Added test boilerplate files
+- Properly serves files when initiating "Create Boilerplate" command into VS Code
+
+## [0.0.1] - 2021-04-19
+
+### Add
+
+- Added README
+- Initialized Yo Code with starter code for VSCode Extension
