+++ src/Boilerplates/pkgjsonBoilerplate.ts
@@ -1,0 +1,16 @@
+const pkgjsonString = `{
+  "name": "upnextapp",
+  "version": "1.0.0",
+  "description": "insert description here!",
+  "main": "index.js",
+  "scripts": {
+    "test": "echo \\"Error: no test specified\\" && exit 1",
+    "dev": "next dev",
+    "build": "next build",
+    "start": "next start"
+  },
+  "author": "yourName",
+  "license": "ISC"
+}`;
+
+export = pkgjsonString;
